+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Misc
+- replace usage of `pandas.DataFrame` with `pyarrow.Table`
+
 ## [0.6.2] - 2025-09-28
 Patch release fixing a bug when registering Arrow Datasets as tables instead of
 views with DuckDB query engine, leading to performance degradation on larger tables.